+++ src/openms_gui/include/OpenMS/VISUAL/MISC/sources.cmake
@@ -3,7 +3,7 @@ set(directory include/OpenMS/VISUAL/MISC)
 
 ### list all header files of the directory here
 set(sources_list_h
-Helpers.h
+GUIHelpers.h
 )
 
 ### add path to the filenames
